backend/core/fsa_reader.py
from Bio import SeqIO
import numpy as np

class FSAReader:
    @staticmethod
    def process_file(content: bytes, filename: str) -> dict:
        """
        Procesa un archivo FSA y extrae los datos básicos
        """
        # Por ahora retornamos datos de ejemplo
        return {
            "filename": filename,
            "channels": 4,
            "data_points": 5000,
            "size_standard": "LIZ-500",
            "alleles": {}
        }
