backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
biopython==1.83
numpy==1.26.2
scipy==1.11.4
pandas==2.1.4
openpyxl==3.1.2
pyinstaller==6.3.0
